+++ extensions/iceworks-material-helper/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.7.1
 
+- feat: compatible block and component material name
 - chore: upgrade `@iceworks/common-service`, `@iceworks/material-engine` and `@iceworks/project-service` version
 
 ## 0.7.0